chore(test): de-duplicate shared fixtures/constants across the test tree
Phase 4 of the test migration: shared OTA-image infrastructure was copied
verbatim across many subtree conftests. Consolidate it to its minimum scope.

Moved to the top-level tests/conftest.py (common, cross-tier, non-autouse):
- container path constants (/ota-image*, /certs*, CSV metadata fixtures),
  HTTP server addr/port/URL constants, KERNEL_PREFIX/INITRD_PREFIX, versions,
  OTA_UPDATER_MODULE
- _get_kernel_version()/KERNEL_VERSION, SlotMeta, SlotAB
- _run_http_server()/_serve_directory() and the legacy/v1 ota-image server
  fixtures (now a single shared session fixture each — also removes a latent
  double-bind on ports 8080/8081 when ota_core e2e and ota_metadata e2e ran in
  the same session)
- ab_slots, ab_slots_for_inplace, resource_dir, fst_db_helper

Moved to tests/_fixtures_ota_core.py (autouse, imported only by the ota_core
conftests so the patch scope stays confined — must NOT be top-level autouse,
which would patch ota_core for all 1110 tests):
- mock_ensure_umount, mock_ensure_mount, mock_certs_dir, mock_fstrim

Subtree conftests now re-export the shared names (so test modules keep
importing via `from .conftest`) and keep only their unique bits (perf report
machinery + dummy boot controller, the real-vs-dummy COOKIES_JSON, the
create_standby verify_resources/ab_slots_for_rebuild, ota_metadata signing
artifacts). Net -276 lines, ~750 duplicated lines removed.

Verified via docker py313 runner: full regular suite 1107 passed / 1 skipped
(470s) plus the 2 performance tests passed; 1110 collected.
