+++ tests/integration/test_otaclient/test_ota_core/conftest.py
@@ -13,45 +13,23 @@
 # limitations under the License.
 """Shared mocks for ota_core integration tests.
 
-`_updater.ensure_umount` shells out to the real OS umount syscall; the
-updater calls it on the session workdir during `execute()` cleanup, so
-integration tests that drive the updater state machine must mock it.
-The CA cert dir is repointed at `/certs` (baked into the test container
-image) so any code path that consults `cfg.CERT_DPATH` works without
-relying on the production install path. `fstrim_at_subprocess` is
-silenced so nothing tries to fstrim during tests.
+`_updater.ensure_umount` shells out to the real OS umount syscall; the updater
+calls it on the session workdir during `execute()` cleanup, so integration
+tests that drive the updater state machine must mock it. The CA cert dir is
+repointed at `/certs` (baked into the test container image) so any code path
+that consults `cfg.CERT_DPATH` works without relying on the production install
+path. `fstrim_at_subprocess` is silenced so nothing tries to fstrim during
+tests.
+
+The fixture bodies are shared with the unit/e2e ota_core subtrees via
+`tests._fixtures_ota_core`; they are imported (not redefined) here so the
+autouse scope stays confined to this subtree.
 """
 
 from __future__ import annotations
 
-import pytest
-import pytest_mock
-
-from otaclient import ota_core
-
-# Path to certs baked into the test docker image (see docker/test_base/Dockerfile).
-_CONTAINER_CERTS_DIR = "/certs"
-
-OTA_UPDATER_MODULE = ota_core._updater.__name__
-
-
-@pytest.fixture(autouse=True)
-def mock_ensure_umount(mocker: pytest_mock.MockerFixture) -> None:
-    mocker.patch(f"{OTA_UPDATER_MODULE}.ensure_umount")
-
-
-@pytest.fixture(autouse=True, scope="module")
-def mock_certs_dir(module_mocker: pytest_mock.MockerFixture):
-    """Repoint CERT_DPATH at the certs baked into the test container image."""
-    from otaclient.configs.cfg import cfg as _cfg
-
-    module_mocker.patch.object(_cfg, "CERT_DPATH", _CONTAINER_CERTS_DIR)
-
-
-@pytest.fixture(autouse=True, scope="module")
-def mock_fstrim(module_mocker: pytest_mock.MockerFixture):
-    """Block real `fstrim` invocations from leaking out of integration tests."""
-    module_mocker.patch(
-        f"{OTA_UPDATER_MODULE}.fstrim_at_subprocess",
-        module_mocker.MagicMock(),
-    )
+from tests._fixtures_ota_core import (  # noqa: F401
+    mock_certs_dir,
+    mock_ensure_umount,
+    mock_fstrim,
+)
